admin: Move Future UI story-only helpers to an unpublished directory
The future-ui `utils/` directory held both runtime helpers that
components import and the package publishes, and helpers only
Storybook uses. The Babel build compiles everything under `src`
into the published `lib`, so the story-only helpers shipped to
consumers with no reason to. Give story-only code its own directory
and exclude it from the build, the way the `cli/` token generator
is already excluded.

- **Build-only exclusions live in `.babelrc.build.json`, not the
  package's `.babelrc.json`.** Storybook's Vite dev server also
  honors `.babelrc.json`'s `ignore`, so excluding the directories
  there would 404 them in Storybook and break every story that
  imports a helper; the build loads both files and unions their
  ignore lists.
