+++ packages/admin/admin/.babelrc.build.json
@@ -1,0 +1,3 @@
+{
+    "ignore": ["src/future-ui/cli/**", "src/future-ui/storybook/**"]
+}
