+++ packages/admin/admin/.babelrc.json
@@ -1,4 +1,4 @@
 {
-    "ignore": ["**/*.test.ts", "**/*.test.tsx", "**/*.spec.ts", "**/*.spec.tsx", "**/__tests__/**", "src/future-ui/cli/**"],
+    "ignore": ["**/*.test.ts", "**/*.test.tsx", "**/*.spec.ts", "**/*.spec.tsx", "**/__tests__/**"],
     "presets": ["@comet/admin-babel-preset"]
 }
