+++ packages/admin/admin/package.json
@@ -26,7 +26,7 @@
     "scripts": {
         "build": "pnpm run clean && run-p build:babel build:types",
         "build-storybook": "storybook build --disable-telemetry",
-        "build:babel": "pnpm exec babel ./src -x \".ts,.tsx\" -d lib",
+        "build:babel": "pnpm exec babel ./src -x \".ts,.tsx\" -d lib --config-file ./.babelrc.build.json",
         "build:types": "tsc --project ./tsconfig.build.json --emitDeclarationOnly",
         "clean": "rimraf lib",
         "generate-future-ui-theme-tokens": "ts-node ./src/future-ui/cli/generateThemeTokens.ts",
@@ -39,7 +39,7 @@
         "lint:prettier": "pnpm exec prettier --check '*.{ts,js,json,md,yml,yaml}'",
         "lint:tsc": "tsc --noEmit",
         "start": "run-p start:babel start:types",
-        "start:babel": "pnpm exec babel ./src -x \".ts,.tsx\" -d lib -w",
+        "start:babel": "pnpm exec babel ./src -x \".ts,.tsx\" -d lib -w --config-file ./.babelrc.build.json",
         "start:types": "tsc --project ./tsconfig.json --emitDeclarationOnly --watch --preserveWatchOutput",
         "storybook": "storybook dev -p 26646 --no-open --disable-telemetry",
         "test": "pnpm run test:unit && pnpm run test:storybook",
