+++ packages/admin/admin/src/future-ui/components/button/__stories__/Button.stories.tsx
@@ -1,7 +1,7 @@
 import { ArrowRight, Favorite } from "@comet/admin-icons";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import { cssVarsProviderDecorator } from "../../../utils/cssVarsProviderDecorator";
+import { cssVarsProviderDecorator } from "../../../storybook/cssVarsProviderDecorator";
 import { Button } from "../Button";
 
 const meta: Meta<typeof Button> = {
